requirements.txt
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
db-dtypes==1.2.0
filelock==3.20.0
fsspec==2025.9.0
ftfy==6.3.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==3.3.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
hf-xet==1.1.10
huggingface-hub==0.36.0
idna==3.10
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
locate==1.1.1
marisa-trie==1.3.1
MarkupSafe==3.0.3
mpmath==1.3.0
msgpack==1.1.2
networkx==3.5
numpy==1.26.4
packaging==25.0
pandas==2.3.1
playwright==1.54.0
praw==7.8.1
pytest==8.3.4
pytest-asyncio==0.25.2
prawcore==2.4.0
proto-plus==1.26.1
protobuf==6.32.0
py4j==0.10.9.9
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyee==13.0.0
pyspark==4.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
rsa==4.9.1
safetensors==0.6.2
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
tokenizers==0.22.1
torch==2.2.2
tqdm==4.67.1
transformers==4.57.1
typing_extensions==4.14.1
tzdata==2025.2
update-checker==0.18.0
urllib3==2.5.0
vaderSentiment==3.3.2
wcwidth==0.2.14
websocket-client==1.9.0
wheel==0.45.1
wordfreq==3.1.1
snscrape==0.7.0.20230622
google-cloud-bigquery==3.38.0
